@@ -4,13 +4,12 @@ pub const GRID_SIZE_X: usize = 96;
 pub const GRID_SIZE_Y: usize = 60;
 pub const VECTOR_SCALE: f64 = 20.0;
 pub const VECTOR_WIDTH: f64 = 2.0;
-pub const CIRCLE_TOLERANCE: f32 = 1.0;
 
-pub const DEFAULT_MOVE_SPEED: f64 = 0.05;
-pub const DEFAULT_NOISE_SCALE: f64 = 0.01;
-pub const DEFAULT_NOISE_SCALE_INCREMENT: f64 = DEFAULT_NOISE_SCALE * 0.2;
-pub const DEFAULT_NOISE_SPEED: f64 = 0.00001;
-pub const DEFAULT_NOISE_SPEED_INCREMENT: f64 = DEFAULT_NOISE_SPEED * 0.2;
+pub const DEFAULT_MOVE_SPEED: f64 = 0.001;
+pub const DEFAULT_NOISE_SCALE: f64 = 0.001;
+pub const DEFAULT_NOISE_SCALE_INCREMENT: f64 = DEFAULT_NOISE_SCALE * 0.01;
+pub const DEFAULT_NOISE_SPEED: f64 = 0.000005;
+pub const DEFAULT_NOISE_SPEED_INCREMENT: f64 = DEFAULT_NOISE_SPEED * 0.01;
 
 pub const SCREEN_W: usize = 1920;
 pub const SCREEN_H: usize = 1080;
