@@ -2,31 +2,31 @@ use log::trace;
 use noise::{Fbm, NoiseFn};
 use rand::{prelude::StdRng, Rng, SeedableRng};
 
-pub struct DomainWarpingNoise {
-    fbm: Fbm,
+pub struct DomainWarpingNoise<T> {
+    fbm: Fbm<T>,
     params: DomainWarpingNoiseParams,
 }
 
-impl DomainWarpingNoise {
+impl<T: Default + noise::Seedable> DomainWarpingNoise<T> {
     pub fn new(params: DomainWarpingNoiseParams) -> Self {
         Self {
-            fbm: Fbm::new(),
+            fbm: Fbm::new(0),
             params,
         }
     }
 
     pub fn inigo() -> Self {
         Self {
-            fbm: Fbm::new(),
+            fbm: Fbm::new(0),
             params: DomainWarpingNoiseParams::inigo(),
         }
     }
 }
 
-impl Default for DomainWarpingNoise {
+impl<T: Default + noise::Seedable> Default for DomainWarpingNoise<T> {
     fn default() -> Self {
         Self {
-            fbm: Fbm::new(),
+            fbm: Fbm::new(0),
             params: Default::default(),
         }
     }
@@ -47,35 +47,35 @@ impl DomainWarpingNoiseParams {
 
     pub fn random() -> Self {
         // rng.gen handles arrays just fine but fills them with numbers between 0.0 and 1.0
-        // gen_range allows for numbers from a larger range but it can't handle arrays :(
-        let mut rng = StdRng::from_entropy();
+        // gen_range allows for numbers from a larger range, but it can't handle arrays :(
+        let mut rng = StdRng::from_rng(&mut rand::rng());
         let q_range = -100.0..100.0;
         let q: [f64; 9] = [
-            rng.gen_range(q_range.clone()),
-            rng.gen_range(q_range.clone()),
-            rng.gen_range(q_range.clone()),
-            rng.gen_range(q_range.clone()),
-            rng.gen_range(q_range.clone()),
-            rng.gen_range(q_range.clone()),
-            rng.gen_range(q_range.clone()),
-            rng.gen_range(q_range.clone()),
-            rng.gen_range(q_range.clone()),
+            rng.random_range(q_range.clone()),
+            rng.random_range(q_range.clone()),
+            rng.random_range(q_range.clone()),
+            rng.random_range(q_range.clone()),
+            rng.random_range(q_range.clone()),
+            rng.random_range(q_range.clone()),
+            rng.random_range(q_range.clone()),
+            rng.random_range(q_range.clone()),
+            rng.random_range(q_range.clone()),
         ];
 
         let r_range = -100.0..100.0;
         let r: [f64; 9] = [
-            rng.gen_range(r_range.clone()),
-            rng.gen_range(r_range.clone()),
-            rng.gen_range(r_range.clone()),
-            rng.gen_range(r_range.clone()),
-            rng.gen_range(r_range.clone()),
-            rng.gen_range(r_range.clone()),
-            rng.gen_range(r_range.clone()),
-            rng.gen_range(r_range.clone()),
-            rng.gen_range(r_range.clone()),
+            rng.random_range(r_range.clone()),
+            rng.random_range(r_range.clone()),
+            rng.random_range(r_range.clone()),
+            rng.random_range(r_range.clone()),
+            rng.random_range(r_range.clone()),
+            rng.random_range(r_range.clone()),
+            rng.random_range(r_range.clone()),
+            rng.random_range(r_range.clone()),
+            rng.random_range(r_range.clone()),
         ];
-        let qn: f64 = rng.gen_range(0.001..5.0);
-        let rn: f64 = rng.gen_range(0.001..5.0);
+        let qn: f64 = rng.random_range(0.001..5.0);
+        let rn: f64 = rng.random_range(0.001..5.0);
 
         let s = Self { q, r, qn, rn };
 
@@ -88,7 +88,7 @@ impl DomainWarpingNoiseParams {
     }
 
     // https://www.iquilezles.org/www/articles/warp/warp.htm
-    fn inigo() -> Self {
+    pub(crate) fn inigo() -> Self {
         Self {
             q: [0.0, 0.0, 0.0, 5.2, 1.3, 0.0, 0.0, 0.0, 0.0],
             r: [1.7, 9.2, 0.0, 8.3, 2.8, 0.0, 0.0, 0.0, 0.0],
@@ -110,7 +110,9 @@ impl Default for DomainWarpingNoiseParams {
 }
 
 // Image for f(p) = fbm(p+fbm(p+fbm(p)))
-impl NoiseFn<[f64; 3]> for DomainWarpingNoise {
+impl<T> NoiseFn<f64, 3> for DomainWarpingNoise<T>
+where
+T: NoiseFn<f64, 3>,{
     fn get(&self, p: [f64; 3]) -> f64 {
         let [q0, q1, q2, q3, q4, q5, q6, q7, q8] = self.params.q;
         let q = vec3(
